Mixins replacement
Moved mixins to a related folders, imported to components
