+++ src/components/user/UserSettings.vue
@@ -39,7 +39,7 @@
 import ImageInput from '@/components/ImageInput.vue'
 import Message from '@/components/Message.vue'
 import CountrySelect from '@/components/CountrySelect.vue'
-import updateUser from '@/mixins/updateUser.js'
+import updateUser from '@/mixins/users/updateUser.js'
 
 export default {
   name: 'user-settings',
